src/nat: relocate inline tests to nat_tests.rs
nat.rs reaches 1,582 LOC with a 1,028-LOC `#[cfg(test)] mod tests` block.
Same playbook as #1080-#1089: extract to a sibling nat_tests.rs.

After: nat.rs drops from 1,582 → 557 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
